+++ com/hbm/inventory/container/ContainerCoreAdvanced.java
@@ -73,7 +73,7 @@ public class ContainerCoreAdvanced extends Container {
 	public void addCraftingToCrafters(ICrafting crafting) {
 		super.addCraftingToCrafters(crafting);
 		crafting.sendProgressBarUpdate(this, 0, this.diFurnace.progress);
-		crafting.sendProgressBarUpdate(this, 1, this.diFurnace.power);
+		crafting.sendProgressBarUpdate(this, 1, (int)this.diFurnace.power);
 	}
 	
 	@Override
@@ -134,12 +134,12 @@ public class ContainerCoreAdvanced extends Container {
 			
 			if(this.power != this.diFurnace.power)
 			{
-				par1.sendProgressBarUpdate(this, 1, this.diFurnace.power);
+				par1.sendProgressBarUpdate(this, 1, (int)this.diFurnace.power);
 			}
 		}
 		
 		this.progress = this.diFurnace.progress;
-		this.power = this.diFurnace.power;
+		this.power = (int)this.diFurnace.power;
 	}
 	
 	@Override
